extract symbols
